+++ src/Xamarin.Auth.Core/AssemblyInfo.cs
