+++ src/Xamarin.Auth.Core/AuthException.cs
