+++ src/Xamarin.Auth.Core/Authenticator.cs
