+++ src/Xamarin.Auth.Core/ExceptionEx.cs
