+++ src/Xamarin.Auth.Core/FormAuthenticator.cs
