+++ src/Xamarin.Auth.Core/OAuth2Authenticator.cs
