+++ src/Xamarin.Auth.Core/Request.cs
