+++ src/Xamarin.Auth.Core/Response.cs
