+++ src/Xamarin.Auth.Core/System.Json/JavaScriptReader.cs
