+++ src/Xamarin.Auth.Core/System.Json/JsonArray.cs
