+++ src/Xamarin.Auth.Core/System.Json/JsonObject.cs
