+++ src/Xamarin.Auth.Core/System.Json/JsonPrimitive.cs
