+++ src/Xamarin.Auth.Core/System.Json/JsonType.cs
