+++ src/Xamarin.Auth.Core/System.Json/JsonValue.cs
