+++ src/Xamarin.Auth.Core/WebAuthenticator.cs
