+++ src/Xamarin.Auth.Core/WebEx.cs
