+++ src/Xamarin.Auth.Core/WebRedirectAuthenticator.cs
