+++ src/Xamarin.Auth.Core/WebUtilities.cs
