+++ src/Xamarin.Auth.Core/Xamarin.Auth.Core.csproj
@@ -34,84 +34,40 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="..\Xamarin.Auth\Account.cs">
-      <Link>Account.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Auth\AccountStore.cs">
-      <Link>AccountStore.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Auth\AssemblyInfo.cs">
-      <Link>AssemblyInfo.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Auth\Authenticator.cs">
-      <Link>Authenticator.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Auth\AuthException.cs">
-      <Link>AuthException.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Auth\DelegatedOAuth2Authenticator.cs">
-      <Link>DelegatedOAuth2Authenticator.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Auth\ExceptionEx.cs">
-      <Link>ExceptionEx.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Auth\FormAuthenticator.cs">
-      <Link>FormAuthenticator.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Auth\OAuth2Authenticator.cs">
-      <Link>OAuth2Authenticator.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Auth\OAuth2Request.cs">
-      <Link>OAuth2Request.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Auth\Request.cs">
-      <Link>Request.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Auth\Response.cs">
-      <Link>Response.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Auth\System.Json\JavaScriptReader.cs">
-      <Link>System.Json\JavaScriptReader.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Auth\System.Json\JsonArray.cs">
-      <Link>System.Json\JsonArray.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Auth\System.Json\JsonObject.cs">
-      <Link>System.Json\JsonObject.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Auth\System.Json\JsonPrimitive.cs">
-      <Link>System.Json\JsonPrimitive.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Auth\System.Json\JsonType.cs">
-      <Link>System.Json\JsonType.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Auth\System.Json\JsonValue.cs">
-      <Link>System.Json\JsonValue.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Auth\WebAuthenticator.cs">
-      <Link>WebAuthenticator.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Auth\WebEx.cs">
-      <Link>WebEx.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Auth\WebRedirectAuthenticator.cs">
-      <Link>WebRedirectAuthenticator.cs</Link>
-    </Compile>
-    <Compile Include="..\Xamarin.Auth\WebUtilities.cs">
-      <Link>WebUtilities.cs</Link>
-    </Compile>
-  </ItemGroup>
-  <ItemGroup>
     <Reference Include="Polenter.SharpSerializer.PCL, Version=1.0.0.0, Culture=neutral, PublicKeyToken=8f4f20011571ee5f, processorArchitecture=MSIL">
       <HintPath>..\packages\SharpSerializer.PCL.2.18.7\lib\portable-net45+sl50+win+wpa81+wp80\Polenter.SharpSerializer.PCL.dll</HintPath>
       <Private>True</Private>
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <None Include="packages.config" />
+    <None Include="project.json" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Properties\" />
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Include="Account.cs" />
+    <Compile Include="AccountStore.cs" />
+    <Compile Include="AssemblyInfo.cs" />
+    <Compile Include="Authenticator.cs" />
+    <Compile Include="AuthException.cs" />
+    <Compile Include="DelegatedOAuth2Authenticator.cs" />
+    <Compile Include="ExceptionEx.cs" />
+    <Compile Include="FormAuthenticator.cs" />
+    <Compile Include="OAuth2Authenticator.cs" />
+    <Compile Include="OAuth2Request.cs" />
+    <Compile Include="Request.cs" />
+    <Compile Include="Response.cs" />
+    <Compile Include="System.Json\JavaScriptReader.cs" />
+    <Compile Include="System.Json\JsonArray.cs" />
+    <Compile Include="System.Json\JsonObject.cs" />
+    <Compile Include="System.Json\JsonPrimitive.cs" />
+    <Compile Include="System.Json\JsonType.cs" />
+    <Compile Include="System.Json\JsonValue.cs" />
+    <Compile Include="WebAuthenticator.cs" />
+    <Compile Include="WebEx.cs" />
+    <Compile Include="WebRedirectAuthenticator.cs" />
+    <Compile Include="WebUtilities.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
