--- src/Xamarin.Auth.Core/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="SharpSerializer.PCL" version="2.18.7" targetFramework="portable45-net45+win8+wpa81" />
-</packages>
